Theory

After assembling the discrete FE operators we are left with large system of ODEs in time, i.e.

\[\mathbf{M} \mathbf{c}'(t) + \mathbf{A} \mathbf{c}(t) = \mathbf{r}(\mathbf{c})\]

where $\mathbf{M}, \mathbf{A}$ and $\mathbf{r}$ represents the assembled mass matrix, diffusion & advection matrices and reaction operator, respectively. In this represenation it should become clear why we are left with a system of ODEs. Furthermore note that the reaction operator is dependent of the current solution of $c$ and thus introduces a non-linearity.

This non-linearity is right now bypassed by an operator splitting and in this particular case by an sequential solving. So, we first solve the nonlinear micro and then use it and solve the linear problem.

Implementation

Catalyst.solve — Function

solve(Dᵢ::Float64, k::Float64, kᵧ::Float64, 
    input_exp::Array, output_exp::Array;
    N=(100,), L=5e-2, w=1.9128e-4 * (1 / 0.37), 
    T = 1000, Δt=1, Dₑ=1e-9, rᵢ=2.15e-7, 
    h= L/N[1], δT = h/(2 * abs(w)),
    progress=true, calibration=false,
    microsave=false, microsave_time = (250, 300, 350, 400),
    microsave_location=((10,1), (50,1), (80,1)),
    microcomp_type=:linear, microsave_path="simulation/",
    Q=0., kₙ=0.,
    micromesh=Parser.getGrid(projectdir("test/catalyst.msh")))

is the main function of the package which starts a FE computation with nested FE computations in material points. Sets up the finite element spaces, discretizes the operators, applies boundary conditions and solves the time dependent problem in a loop.

Arguments

Dᵢ::Float64: is the microscopic diffusion.

k::Float64, kᵧ::Float64, Q::Float64, kₙ::Float64: are microscopic parameters.

input_exp::Array, output_exp::Array: are the experiment measurements.

progress::Bool: enables/disables a progress bar for the macroscopic time steps.

calibration::Bool: enables/disables a returned error between the last node concentration and the output experiment concentration.

microsave::Bool, microsave_time::Tuple and microsave_location::Tuple: controls which times and locations of the microscopic problems are saved to the disk

microcomp_type::Symbol: decides whether or not linear or nonlinear micro computations are done.

microsave_path::String: path where to store the micro .pvd files inside data/ dir

micromesh::JuAFEM.Grid: describes the microscopic domain.

Returns either two arrays, one dimensional concentration field c at each time step and the assembled reaction operator at each time step or returns the squarred error (scalar).