Structs

Suitable datastructures are developed that hold as their fields all information about the microstructure. In case of the simplified, linerized sink term CatalystStateODE can be used to store various numbers associated with the catalyst state. When a PDE description of the microscale is used CatalystStatePDE should be used. This struct assembles by a given micro diffusion the corresponding linear operators of the micro scale and optionally assembles in each catalyst_update! call the nonlinear parts if :nonlinear is passed as the microcomp_type.

Catalyst.CatalystStateODE — Type

CatalystStateODE(D_i, kᵧ, k, h, r, V, A, Δt, coeff, cᵧ_old)

instantiates a CatalystStateODE struct holding all necessary values for the simplified microscale formulation

Catalyst.CatalystStatePDE — Type

CatalystStatePDE(D_i, k_γ, mesh, Q, kₙ)

constructs a CatalystStatePDE struct that stores all information about the microstructure. Therefore, assembles also the linear diffusion and mass matrices that are stored in M, K and their associated sum in A=K+k_γ*M

Assembly

Most of the assembly is handled while constructing the corresponding CatalystState struct. However, in case of a nonlinear computation the nonlinear parts, i.e. the Jacobian of the residual has to be assembled in each newton iteration, due to its nonlinear nature.

Catalyst.assemble_nonlinear_micro_global! — Function

function assemble_nonlinear_micro_global!(K::SparseMatrixCSC{Float64,Int64}, f::Array{Float64,1}, dh::DofHandler, cv::CellScalarValues, c::Array{Float64,1}, Δt, D, Q, kₙ, cⁿ, 𝐀::SparseMatrixCSC{Float64,Int64})

Assembles only the nonlinear part of the jacobian, so needs to add the linear part after nonlinear assemble, i.e. assemble jacobi K, add mass matrix M and Diffusion Matrix Catalyst.K (𝐀) on top

Care the interface

Catalyst.assemble_nonlinear_micro_global! is the only assembly routine which mutates the input arguments Jacobi $\mathbf{K}$ and residual $\mathbf{f}$

Catalyst.assemble_nonlinear_micro_element! — Function

function assemble_nonlinear_micro_element!(ke, ge, cell, cv, ce, Δt, D, Q, kₙ, cⁿₑ)

assembles the element jacobi for the newton iteration. This function is never called by any user, it will be called by assemble_nonlinear_micro_global!

Micro Solver

From a user perspective only catalyst_update is invoked which determines based on the input which micro computation need to be done. In case of an array of arrays holding CatalystStateODEs the update is done based on a suitable ODE.

If a array of array of CatalystStatePDE is passed, either a linear or nonlinear micro PDE is solved.

Catalyst.catalyst_update! — Function

catalyst_update!( cellvalues::CellScalarValues{dim}, dh::DofHandler, c::AbstractVector, Catalysts::Array{Array{CatalystStateODE,1},1}, δT::Float64, w::Float64)

updates all CatalystStateODE structs that need to be passed as a arrays of arrays. The first array corresponds to the element index and in each element index there is a nested array for all gauss points holding in each entry a CatalystStateODE

The function then updates the state by the corresponding ordinary differnetial equation

catalyst_update!(cellvalues::CellScalarValues{dim}, dh::DofHandler, c::AbstractVector, Catalysts::Array{Array{CatalystStatePDE,1},1}, t::Number, computation_type::Symbol)

updates all CatalystStatePDE structs that need to be passed as a arrays of arrays. The first array corresponds to the element index and in each element index there is a nested array for all gauss points holding in each entry a CatalystStatePDE

The function then updates the state by the corresponding partial differnetial equation. the variable computation_type can either be :linear or :nonlinear and thereby determines if a linear or nonlinear PDE is solved. In case of the nonlinear PDE the nonlinearity is introduced by a source/sink term and is in this case the langmuir isotherm formulation. However can be changed without any big hurdles

Linear Micro Solver

Catalyst.microcomputation_linear! — Function

micrcomputation_linear!!(cₑ::Float64, Catalyst::CatalystStatePDE)

solves the discretized linear finite element problem with the current macroscopic concentration cₑ as the value for the Dirichlet boundary condition. After solving the linear system the previous concentration of the Catalyst is updated to the current solution.

Besides updating the previous concentration, this function also updates the current flux across the boundary Catalyst.cᵧ by looping over all boundary faces and their corresponding gauss points and evaluates there the sum

\[c_{\Gamma} = \int_{\partial \Omega}(\mathbf{D} \cdot \nabla c)\cdot \mathbf{n}\ dA\]

Nonlinear Micro Solver

Catalyst.microcomputation_nonlinear! — Function

micrcomputation_nonlinear!!(cₑ::Float64, Catalyst::CatalystStatePDE)

solves the discretized nonlinear finite element problem with the current macroscopic concentration cₑ as the value for the Dirichlet boundary condition. After setting the ConstraintHandler up the nonlinear parts are assembled by assemble_nonlinear_micro_global! and assemble_nonlinear_micro_element!, respectively, within a Newton Iteration loop.

As soon as the solution of the current time step is found the very same flux across the boundary is computed as in microcomputation_linear!.

Micro Source/Sink

Catalyst.langmuir_isotherm′ — Function

langmuir_isotherm′(c¯, Q, kₙ)

computes the first derivative w.r.t. c¯ of the langmuir isotherm formulation, where c¯ is the current Newton guess, Q is accordingly to wiki the value that forms the asymptote, kₙ is the Langmuir-Sorptioncoefficient. Returns a scalar.

\[f'(c^-, Q, k_n) = Q\ k_n\ (1+k_n\ c^-)^{-2}\]

Catalyst.langmuir_isotherm″ — Function

langmuir_isotherm″(c¯, Q, kₙ)

computes the second derivative w.r.t. c¯ of the langmuir isotherm formulation.

\[f''(c^-, Q, k_n) = -2Q\ k_n^2\ (1+k_n\ c^-)^{-3}\]