Visualization

Catalyst.plotAnimation — Function

plotAnimation(storage::Array, gifname::String, lim=(0,1))

Plots a collection of one dimensional arrays to a gif. Applicable to plot the Reaction Operator or the one dimensional concentration field at each time instance over the domain. Both mentioned things are returned by Catalyst.solve(...)

Catalyst.plotOverTime — Function

plotOverTime(storage::Array, ylim=(0,1), save=false, figname=missing)

takes the last entry of the storage array in each time step and plots it over time. Useful to plot the concentration field returned by Catalyst.solve() over time.

I/O

Catalyst.save1D — Function

save1D(storage::Array, savename::String, path::String; headername="data")

writes a one dimensional field to disk by providing the field in storage, a savename as String and a path.